move h3 tests
